Handle user streams and args in subscribers
